SSIS-224 - Ajuste para que se valide y guarde el campo de disciplinas solo cuando se selecciona monitor

+++ app/Repositories/UserRepository.php
@@ -108,11 +108,13 @@ class UserRepository
                 }
 
                 // Diciplinas - usuarios
-                foreach ($user['disciplines'] as $key => $value) {
-                    DisciplineUser::create([
-                        'user_id' => $new_user->id,
-                        'disciplines_id' => $value,
-                    ]);
+                if($user['roles'] == '12'){
+                    foreach ($user['disciplines'] as $key => $value) {
+                        DisciplineUser::create([
+                            'user_id' => $new_user->id,
+                            'disciplines_id' => $value,
+                        ]);
+                    }
                 }
             }else{
                 // Roles - usuarios
@@ -219,7 +221,7 @@ class UserRepository
                 }
 
                 // Diciplinas
-                if($data['disciplines']){
+                if($data['disciplines'] && $data['roles'] == '12'){
                     $discipline = DisciplineUser::where('user_id', $user_up->id)->delete();
                     foreach ($data['disciplines'] as $key => $value) {
                         $discipline = new DisciplineUser();
